[lsp] Disable full text sync.
It's very slow to fetch the entire document, serialize it to JSON and send
to language server. I believe that most useful language servers use the
incremental sync anyway. Later I can revisit this and optimize it.

The only language server I know that uses full text sync is SCry, for
Crystal, but I only used SCry and Solargraph :-P.