Add UI for file reloading when externally modified.
The original plan was to only show a passive banner when the current open view is
externally modified, so the user could just press F5/Esc to decide what to do.

Views not open are asked for a reload only when the user changes to it.

Missing things from original plan:

 - If the current open/focused file is modified the dialog is show, disrrupting the user.
 - There's no indicator on UI open files list to let the user know there are
   external changes in other open files.

All this very easy to do, but moved to 0.2.0.