Using GtkSourcceView.
I decided to put GtkSourceView here and work on the basic things
of the editor UI before finish Tijolo own editor widget using
tree-sitter, so I can release this, start using it and polish the UI
while I keep working on the PieceTable and Code widget.

Meanwhile the editor is supposed to be always working with both
editors, to compile it to use the experimental code widget use the
-Dexperimental flag.

Once the new code widget is ready I will remove the GtkSourceView code
from here, so... no GtkSOurceView types must be exposed outside of
files under `src/gsv` directory.