Fix issue with GTK 4.12.5 that prevent buffer to save.
Just set the buffer as not-modified after load it would solve this.
However I took the oportunity to remove pieces of code from the time
when I was developing a text widget.

Nowadays I plan to go into this adventure of writing a text widget
only after I finish the whole UI around it.