Implement local symbol locator.
This is more to test the API while having something useful at the same time,
so for Ruby and Crystal it uses a very lame code model that just run
some regexp in the files to identify methods, macros and spec tests.

In a near future I want this to be backed by tree-sitter parsers, maybe
discarding LSP support if it ends up good enough.